+++ src/sdk/gui/WListView.java
@@ -92,7 +92,7 @@ public class WListView extends Widget
 	}
 
 	@Override
-	public ClickResponse MouseDown(Object sender, int xp, int yp, byte buttons)
+	public ClickResponse MouseDown(Object sender, int xp, int yp, int buttons)
 	{
 		if (buttons == 1)
 		{
